+++ src/models/incidentals/deduction_type.ts
@@ -11,4 +11,9 @@ export enum DeductionType {
    * The costs get split up between each apartment
    */
   PerApartment = 'Pro Wohnung',
+
+  /**
+   * The incidentals can't be deducted in the invoice
+   */
+  NotDeductable = 'Nicht abrechenbar',
 }